Add Papuh Live
Live Wallpaper created from walls based on user selections
All walls can be managed in the Papuh Live settings, as well as the update interval
If the update interval is set to 30 mins, the wall changes every 30 minutes

We have to use canvas for drawing the walls because the WallpaperManager destroyes the live wall engine
Some corrections are needed for the cropping/resizing/scaling of the wall. Some walls get zoomed in too much on the homescreen.